primer mantenimiento completado

+++ src/app/Components/Activos-Fijos/view-activos/view-activos.component.ts
@@ -1,8 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ActivosFijosService } from '../../../Services/activos-fijos.service';
 import { ActivosFijo } from '../../../model/ActivoFijo';
 import { global } from '../../../Services/Global';
+import Swal from 'sweetalert2';
+import { AddActivosComponent } from '../add-activos/add-activos.component';
+import { formatDate } from '@angular/common';
 @Component({
   selector: 'app-activos',
   templateUrl: './view-activos.component.html',
@@ -13,6 +16,8 @@ export class ActivosComponent implements OnInit {
 
   public url: any;
   public activos: Array<ActivosFijo>;
+  public activo: ActivosFijo;
+  @ViewChild(AddActivosComponent) addmodal: AddActivosComponent;
 
   constructor(
     private _router: Router,
@@ -26,16 +31,50 @@ export class ActivosComponent implements OnInit {
   }
 
   delete(id:number){
-
-    this._activosService.deleteActivos(id).subscribe({
-      next:(res)=>{
-        this.getActivos();
-
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success ml-3',
+        cancelButton: 'btn btn-danger'
       },
-      error:(err)=>{
-        console.log(err);
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Estas seguro?',
+      text: "Esta accion no se puede revertir!",
+      icon: 'warning',
+      showCancelButton: true, 
+      confirmButtonText: 'Si, borrar!',
+      cancelButtonText: 'No, cancelar!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        
+        this._activosService.deleteActivos(id).subscribe({
+          next:(res)=>{
+            this.getActivos();
+            swalWithBootstrapButtons.fire(
+              'Eliminado!',
+              'Este registro fue eliminado correctamente!',
+              'success'
+            )
+          },
+          error:(err)=>{
+            console.log(err);
+          }
+        })
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelado',
+          'Accion cancelada',
+          'error'
+        )
       }
     })
+    
 
   }
 
@@ -52,4 +91,37 @@ export class ActivosComponent implements OnInit {
       }
   })
   }
+
+  CreateMode(){ 
+    this.addmodal.inView=false;
+    this.addmodal.form.enable();
+    this.addmodal.Clear();
+  }
+
+  getActivoById(id:number,mode:number) {
+    this._activosService.getActivos().subscribe({
+      next:(res)=>{
+        if (res.dataList
+        ) {
+            this.activo = res.dataList.find((x:any)=> x.activoFijoId == id)
+            this.addmodal.form.patchValue({
+              ...this.activo, fechaRegistro: formatDate(this.activo.fechaRegistro,'yyyy-MM-dd','en')            
+            });           
+          if(mode == 2){      
+            this.addmodal.form.disable();
+            this.addmodal.inView=true;
+
+          }else{
+            this.addmodal.inView=false;
+            this.addmodal.form.enable();
+          }
+
+        }
+      },
+      error:(err)=>{
+        console.log(err);
+      }
+  })
+  }
+
 }
